refactor(inquiry): moves context into hooks dir
