+++ src/v2/Components/Inquiry/Views/InquiryHowCanWeHelp.tsx
@@ -1,6 +1,6 @@
 import { Button, Text } from "@artsy/palette"
 import React from "react"
-import { useInquiryContext } from "../InquiryContext"
+import { useInquiryContext } from "../Hooks/useInquiryContext"
 
 export const InquiryHowCanWeHelp: React.FC = () => {
   const { next } = useInquiryContext()
