+++ src/v2/Components/Inquiry/Views/InquirySignUp.tsx
@@ -4,7 +4,7 @@ import { useState } from "react"
 import { createRelaySSREnvironment } from "v2/System/Relay/createRelaySSREnvironment"
 import { ReCaptchaContainer } from "v2/Utils/ReCaptchaContainer"
 import { wait } from "v2/Utils/wait"
-import { useInquiryContext } from "../InquiryContext"
+import { useInquiryContext } from "../Hooks/useInquiryContext"
 import { useArtworkInquiryRequest } from "../Hooks/useArtworkInquiryRequest"
 import { signUp } from "v2/Utils/auth"
 import { logger } from "../util"
