+++ src/v2/Components/Inquiry/__tests__/Views/InquirySignUp.jest.tsx
@@ -4,12 +4,12 @@ import { flushPromiseQueue } from "v2/DevTools"
 import { InquirySignUp } from "../../Views/InquirySignUp"
 import { useArtworkInquiryRequest } from "../../Hooks/useArtworkInquiryRequest"
 import { signUp } from "v2/Utils/auth"
-import { useInquiryContext } from "../../InquiryContext"
+import { useInquiryContext } from "../../Hooks/useInquiryContext"
 import { fill } from "../util"
 
 jest.mock("v2/Utils/auth")
 jest.mock("../../Hooks/useArtworkInquiryRequest")
-jest.mock("../../InquiryContext")
+jest.mock("../../Hooks/useInquiryContext")
 jest.mock("v2/Utils/wait", () => ({ wait: () => Promise.resolve() }))
 
 describe("InquirySignUp", () => {
